Add support for Keyboardio Model01 and Model100
This commit adds support for the Keyboardio Model 01 and the Keyboardio
Model 100. These are using the Kaleidoscope firmware and their keymaps
are defined in .ino files. However the format of the keymap definitions
is compatible with the keymap definitions in QMK, therefore they are
supported out of the box by keymapviz.

There was only 1 change to enhance the most common keys by removing the
“Key_” prefix that is used by Kaleidoscope additionally to the “KC_”
prefix used by QMK.

As QMK already provides support for the Model 01 this was added as
“model01_qmk” as well. This was necessary as the keys are specified in a
different order in QMK compared to Kaleidoscope.